x/cli: Split command kinds in different files
